REFACTOR: fix eslint warnings and restore the hot reload
